+++ CHANGELOG.rst
@@ -2,7 +2,7 @@
 Changelog
 =========
 
-0.48.1 (unreleased)
+0.48.2 (unreleased)
 ===================
 
 * fix: Do not put toolbar items into the CKEditor 5 toolbar that have no
@@ -18,6 +18,12 @@ Changelog
   for the active Django language.
 * feat: Allow ``shouldNotGroupWhenFull`` in ``TEXT_EDITOR_SETTINGS`` and accept
   the ``{"items": [...]}`` object form of a toolbar configuration.
+
+0.48.1 (2026-09-14)
+===================
+
+* feat: Honor djangocms-text's ``bodyClass`` setting by applying the class names
+  to the editing root element
 
 0.48.0 (2026-04-26)